Move logging and debug from __init__.py to pydicom.config, for issue #168.
