+++ pydicom/__init__.py
@@ -1,7 +1,7 @@
 # __init__.py for Dicom package
 """pydicom package -- easily handle DICOM files. See Quick Start below.
 
-Copyright (c) 2008-2012 Darcy Mason
+Copyright (c) 2008-2014 Darcy Mason
 This file is part of pydicom, released under a modified MIT license.
    See the file license.txt included with this distribution, also
    available at https://github.com/darcymason/pydicom
@@ -10,8 +10,8 @@ This file is part of pydicom, released under a modified MIT license.
 Quick Start
 -----------
 1. A simple program to read a dicom file, modify a value, and write to a new file::
-    import pydicom
-    dataset = pydicom.read_file("file1.dcm")
+    from pydicom import dicomio
+    dataset = dicomio.read_file("file1.dcm")
     dataset.PatientName = 'anonymous'
     dataset.save_as("file2.dcm")
 
@@ -33,39 +33,6 @@ if sys.version_info < (2, 6, 0):
     raise ImportError("pydicom > 0.9.7 requires python 2.6 or later")
 
 
-# Set up logging system for the whole package.
-# In each module, set logger=logging.getLogger('pydicom')  and the same instance
-#     will be used by all
-# At command line, turn on debugging for all pydicom functions with:
-#        import pydicom
-#        pydicom.debug()
-#  Turn off debugging with
-#       pydicom.debug(False)
-import logging
-
-
-def debug(debug_on=True):
-    """Turn debugging of DICOM file reading and writing on or off.
-    When debugging is on, file location and details about the elements read at
-    that location are logged to the 'pydicom' logger using python's logging module.
-
-    :param debug_on: True (default) to turn on debugging, False to turn off.
-    """
-    global logger, debugging
-    if debug_on:
-        logger.setLevel(logging.DEBUG)
-        debugging = True
-    else:
-        logger.setLevel(logging.WARNING)
-        debugging = False
-
-logger = logging.getLogger('pydicom')
-handler = logging.StreamHandler()
-# formatter = logging.Formatter("%(asctime)s %(levelname)s: %(message)s", "%Y-%m-%d %H:%M") #'%(asctime)s %(levelname)s %(message)s'
-formatter = logging.Formatter("%(message)s")
-handler.setFormatter(formatter)
-logger.addHandler(handler)
-debug(False)  # force level=WARNING, in case logging default is set differently (issue 102)
 
 # pre-pydicom 1.0, read_file and write_file were imported here.
 # Continue to do so, but with deprecation warning
