+++ pydicom/config.py
@@ -7,9 +7,12 @@
 
 # doc strings following items are picked up by sphinx for documentation
 
-# Set the type used to hold DS values
 
-use_DS_decimal = False  # default False; was decimal-based in pydicom 0.9.7
+import logging
+
+# Set the type used to hold DS values
+#    default False; was decimal-based in pydicom 0.9.7
+use_DS_decimal = False  
 
 
 def DS_decimal(use_Decimal_boolean=True):
@@ -24,7 +27,7 @@ def DS_decimal(use_Decimal_boolean=True):
     else:
         pydicom.valuerep.DSclass = pydicom.valuerep.DSfloat
 
-
+# Configuration flags
 allow_DS_float = False
 """Set allow_float to True to allow DSdecimal instances to be created with floats;
 otherwise, they must be explicitly converted to strings, with the user
@@ -34,3 +37,30 @@ enforce_valid_values = True
 """Raise errors if any value is not allowed by DICOM standard, e.g. DS strings
 that are longer than 16 characters; IS strings outside the allowed range.
 """
+
+
+# Logging system and debug function to change logging level
+logger = logging.getLogger('pydicom')
+handler = logging.StreamHandler()
+formatter = logging.Formatter("%(message)s")
+handler.setFormatter(formatter)
+logger.addHandler(handler)
+
+def debug(debug_on=True):
+    """Turn debugging of DICOM file reading and writing on or off.
+    When debugging is on, file location and details about the elements read at
+    that location are logged to the 'pydicom' logger using python's logging module.
+
+    :param debug_on: True (default) to turn on debugging, False to turn off.
+    """
+    global logger, debugging
+    if debug_on:
+        logger.setLevel(logging.DEBUG)
+        debugging = True
+    else:
+        logger.setLevel(logging.WARNING)
+        debugging = False
+
+# force level=WARNING, in case logging default is set differently (issue 103)
+debug(False)  
+
