+++ pydicom/filereader.py
@@ -10,7 +10,7 @@ import os.path
 import warnings
 import zlib
 from io import BytesIO
-import logging
+
 from pydicom.tag import TupleTag
 from pydicom.dataelem import RawDataElement
 from pydicom.util.hexutil import bytes2hex
@@ -19,7 +19,8 @@ from pydicom.charset import default_encoding, convert_encodings
 from pydicom.compat import in_py2
 from pydicom import compat
 
-logger = logging.getLogger('pydicom')
+from pydicom.config import logger
+import pydicom.config
 
 stat_available = True
 try:
@@ -180,7 +181,7 @@ def data_element_generator(fp, is_implicit_VR, is_little_endian,
     fp_read = fp.read
     fp_tell = fp.tell
     logger_debug = logger.debug
-    debugging = pydicom.debugging
+    debugging = pydicom.config.debugging
     element_struct_unpack = element_struct.unpack
 
     while True:
@@ -439,7 +440,7 @@ def _read_file_meta_info(fp):
 
     # Get group length data element, whose value is the length of the meta_info
     fp_save = fp.tell()  # in case need to rewind
-    debugging = pydicom.debugging
+    debugging = pydicom.config.debugging
     if debugging:
         logger.debug("Try to read group length info...")
     bytes_read = fp.read(8)
@@ -538,7 +539,7 @@ def read_preamble(fp, force):
     """
     logger.debug("Reading preamble...")
     preamble = fp.read(0x80)
-    if pydicom.debugging:
+    if pydicom.config.debugging:
         sample = bytes2hex(preamble[:8]) + "..." + bytes2hex(preamble[-8:])
         logger.debug("{0:08x}: {1}".format(fp.tell() - 0x80, sample))
     magic = fp.read(4)
@@ -697,7 +698,7 @@ def read_file(fp, defer_size=None, stop_before_pixels=False, force=False):
         logger.debug(u"Reading file '{0}'".format(fp))
         fp = open(fp, 'rb')
 
-    if pydicom.debugging:
+    if pydicom.config.debugging:
         logger.debug("\n" + "-" * 80)
         logger.debug("Call to read_file()")
         msg = ("filename:'%s', defer_size='%s', "
